refactor[api-server]: LotteryWinnerService 의 구현체 생성
