+++ src/test/java/com/watermelon/server/lottery/service/LotteryServiceImplTest.java
@@ -36,93 +36,6 @@ class LotteryServiceImplTest {
     private LotteryServiceImpl lotteryService;
 
     @Test
-    @DisplayName("레포지토리에서 반환된 값을 ResponseLotteryWinnerDto 로 감싸서 반환한다.")
-    void getLotteryWinners() {
-
-        //given
-        List<ResponseLotteryWinnerDto> expected = List.of(
-                ResponseLotteryWinnerDto.from("email1@email.com", 1)
-        );
-
-        Mockito.when(lotteryApplierRepository.findByLotteryRankNot(-1)).thenReturn(
-                List.of(
-                        LotteryApplier.builder()
-                                .lotteryRank(1)
-                                .email("email1@email.com")
-                                .build()
-                )
-        );
-
-        //when
-        List<ResponseLotteryWinnerDto> actual = lotteryService.getLotteryWinners();
-
-        //then
-        assertThat(actual).isEqualTo(expected);
-
-    }
-
-    @Test
-    @DisplayName("레포지토리에서 반환된 Participant 엔티티를 ResponseLotteryWinnerInfoDto 로 감싸서 반환한다.")
-    void getLotteryWinnerInfo() {
-
-        //given
-        final LotteryApplier lotteryApplier = LotteryApplier.builder()
-                .name(TEST_NAME)
-                .address(TEST_ADDRESS)
-                .phoneNumber(TEST_PHONE_NUMBER)
-                .uid(TEST_UID)
-                .build();
-
-        final ResponseLotteryWinnerInfoDto responseLotteryWinnerInfoDto = ResponseLotteryWinnerInfoDto.from(lotteryApplier);
-
-        Mockito.when(lotteryApplierRepository.findByUid(TEST_UID)).thenReturn(Optional.of(lotteryApplier));
-
-        //when
-        ResponseLotteryWinnerInfoDto actual = lotteryService.getLotteryWinnerInfo(TEST_UID);
-
-        //then
-        assertThat(actual).isEqualTo(responseLotteryWinnerInfoDto);
-
-    }
-
-    @Test
-    @DisplayName("uid 에 해당하는 참가자 정보를 변경하고 저장한다.")
-    void createLotteryWinnerInfoSuccess(){
-        //given
-        LotteryApplier lotteryApplier = LotteryApplier.builder().uid(TEST_UID).build();
-
-        Mockito.when(lotteryApplierRepository.findByUid(TEST_UID)).thenReturn(
-                Optional.of(lotteryApplier));
-
-        RequestLotteryWinnerInfoDto requestLotteryWinnerInfoDto = RequestLotteryWinnerInfoDto.builder()
-                .phoneNumber(TEST_PHONE_NUMBER)
-                .address(TEST_ADDRESS)
-                .name(TEST_NAME)
-                .build();
-
-        //when
-        lotteryService.createLotteryWinnerInfo(TEST_UID, requestLotteryWinnerInfoDto);
-
-        //then
-        assertThat(lotteryApplier.getPhoneNumber()).isEqualTo(TEST_PHONE_NUMBER);
-        assertThat(lotteryApplier.getAddress()).isEqualTo(TEST_ADDRESS);
-        assertThat(lotteryApplier.getName()).isEqualTo(TEST_NAME);
-        verify(lotteryApplierRepository).save(lotteryApplier);
-
-    }
-
-    @Test
-    @DisplayName("uid 가 없으면 예외를 발생시킨다.")
-    void createLotteryWinnerInfoFailure() {
-        //given
-        Mockito.when(lotteryApplierRepository.findByUid(TEST_UID)).thenReturn(Optional.empty());
-
-        //then
-        assertThatThrownBy(()->lotteryService.createLotteryWinnerInfo(TEST_UID, RequestLotteryWinnerInfoDto.builder().build()))
-                .isInstanceOf(NoSuchElementException.class);
-    }
-
-    @Test
     @DisplayName("uid 에 해당하는 참가자가 있으면, rank 와 참가여부를 반환한다.")
     void getLotteryRankPresentationCase() {
